Move RPrint.code from RCore to RUtil
