+++ libr/core/cmd_print.c
@@ -42,6 +42,31 @@ static const char *help_msg_pp[] = {
 	NULL
 };
 
+static const char *help_msg_pc[] = {
+	"Usage:", "pc", " # Print in code",
+	"pc",  "", "C",
+	"pc*", "", "print 'wx' r2 commands",
+	"pcA", "", ".bytes with instructions in comments",
+	"pca", "", "GAS .byte blob",
+	"pci", "", "C array of bytes with instructions",
+	"pcd", "", "C dwords (8 byte)",
+	"pch", "", "C half-words (2 byte)",
+	"pcJ", "", "javascript",
+	"pcj", "", "json",
+	"pck", "", "kotlin",
+	"pco", "", "Objective-C",
+	"pcp", "", "python",
+	"pcr", "", "rust",
+	"pcS", "", "shellscript that reconstructs the bin",
+	"pcs", "", "string",
+	"pcv", "", "JaVa",
+	"pcV", "", "V (vlang.io)",
+	"pcw", "", "C words (4 byte)",
+	"pcy", "", "yara",
+	"pcz", "", "Swift",
+	NULL
+};
+
 static const char *help_msg_p6[] = {
 	"Usage: p6[de]", "[len]", "base64 decoding/encoding",
 	"p6d", "[len]", "decode base64",
@@ -275,31 +300,6 @@ static const char *help_msg_p_minus[] = {
 	"p-e", "", "show ascii-art bar of entropy per block",
 	"p-h", "", "show histogram analysis of metadata per block",
 	"p-j", "", "show json format",
-	NULL
-};
-
-static const char *help_msg_pc[] = {
-	"Usage:", "pc", " # Print in code",
-	"pc",  "", "C",
-	"pc*", "", "print 'wx' r2 commands",
-	"pcA", "", ".bytes with instructions in comments",
-	"pca", "", "GAS .byte blob",
-	"pci", "", "C array of bytes with instructions",
-	"pcd", "", "C dwords (8 byte)",
-	"pch", "", "C half-words (2 byte)",
-	"pcJ", "", "javascript",
-	"pcj", "", "json",
-	"pck", "", "kotlin",
-	"pco", "", "Objective-C",
-	"pcp", "", "python",
-	"pcr", "", "rust",
-	"pcS", "", "shellscript that reconstructs the bin",
-	"pcs", "", "string",
-	"pcv", "", "JaVa",
-	"pcV", "", "V (vlang.io)",
-	"pcw", "", "C words (4 byte)",
-	"pcy", "", "yara",
-	"pcz", "", "Swift",
 	NULL
 };
 
@@ -4340,294 +4340,6 @@ static inline int cmd_pxb_k(const ut8 *buffer, int x) {
 	return buffer[3 - x] << (8 * x);
 }
 
-static const char* bits_to_c_code_fmtstr(int bits) {
-	switch (bits) {
-	case 16:
-		return "0x%04x";
-	case 32:
-		return "0x%08xU";
-	case 64:
-		return "0x%016" PFMT64x "ULL";
-	default:
-		return "0x%02x";
-	}
-}
-
-static void print_c_instructions(RPrint *p, ut64 addr, const ut8 *buf, int len, int ws) {
-	size_t i;
-
-	ws = R_MAX (1, R_MIN (ws, 8));
-	int bits = ws * 8;
-	const char *fmtstr = bits_to_c_code_fmtstr (bits);
-	len /= ws;
-
-	p->cb_printf ("#define _BUFFER_SIZE %d\n", len);
-	p->cb_printf ("const uint%d_t buffer[_BUFFER_SIZE] = {\n", bits);
-	p->interrupt = false;
-
-	int oleft, left = 0;
-	const int orig_align = p->coreb.cfggeti (p->coreb.core, "asm.cmt.col") - 40;
-	int align = orig_align;
-	int li = 0;
-	char *instr = NULL;
-	for (i = 0; !p->interrupt && i <= len; i++) {
-		if (left == 0) {
-			int pleft = oleft;
-			ut64 at = addr + i;
-			char *is = p->coreb.cmdstrf (p->coreb.core, "ao @ 0x%08"PFMT64x"~^size[1]", at);
-			oleft = left = atoi (is);
-			li = i;
-			free (is);
-			if (instr) {
-				align = orig_align - ((pleft - 1) * 6);
-				while (align-- > 0) {
-					p->cb_printf (" ");
-				}
-				p->cb_printf (" /* %s */\n", instr);
-				free (instr);
-				if (i == len) {
-					break; // return;
-				}
-			}
-			instr = p->coreb.cmdstrf (p->coreb.core, "pi 1 @ 0x%08"PFMT64x, at);
-			r_str_trim (instr);
-		}
-		if (i == len) {
-			break;
-		}
-		if (left == oleft) {
-			p->cb_printf (" ");
-		}
-		r_print_cursor (p, i, 1, 1);
-		p->cb_printf (fmtstr, r_read_ble (buf, p->big_endian, bits));
-		r_print_cursor (p, i, 1, 0);
-		p->cb_printf (", ");
-		buf += ws;
-		left --;
-	}
-	if (left > 0 && left != oleft) {
-		align = orig_align - ((oleft - left - 1) * 6);
-		while (align-- > 0) {
-			p->cb_printf (" ");
-		}
-		p->cb_printf (" /* invalid */");
-	}
-	if (left != oleft) {
-		p->cb_printf ("\n");
-	}
-	p->cb_printf ("};\n");
-}
-
-static void print_c_code(RPrint *p, ut64 addr, const ut8 *buf, int len, int ws, int w) {
-	size_t i;
-
-	ws = R_MAX (1, R_MIN (ws, 8));
-	int bits = ws * 8;
-	const char *fmtstr = bits_to_c_code_fmtstr (bits);
-	len /= ws;
-
-	p->cb_printf ("#define _BUFFER_SIZE %d\n", len);
-	p->cb_printf ("const uint%d_t buffer[_BUFFER_SIZE] = {", bits);
-
-	p->interrupt = 0;
-
-	for (i = 0; !p->interrupt && i < len; i++) {
-		if (!(i % w)) {
-			p->cb_printf ("\n  ");
-		}
-		r_print_cursor (p, i, 1, 1);
-		p->cb_printf (fmtstr, r_read_ble (buf, p->big_endian, bits));
-		if ((i + 1) < len) {
-			p->cb_printf (",");
-
-			if ((i + 1) % w) {
-				p->cb_printf (" ");
-			}
-		}
-		r_print_cursor (p, i, 1, 0);
-		buf += ws;
-	}
-	p->cb_printf ("\n};\n");
-}
-
-R_API void r_print_code(RPrint *p, ut64 addr, const ut8 *buf, int len, char lang) {
-	int i, w = (int)(p->cols * 0.7);
-	if (w < 1) {
-		w = 1;
-	}
-	switch (lang) {
-	case '?':
-		r_core_cmd_help ((RCore *)p->user, help_msg_pc);
-		break;
-	case '*':
-		p->cb_printf ("wx ");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			if (i && !(i % 16)) {
-				p->cb_printf (";s+16\nwx ");
-			}
-			p->cb_printf ("%02x", buf[i]);
-		}
-		if (i && !(i % 16)) {
-			p->cb_printf (";s+16\n");
-		} else {
-			p->cb_printf (";s+%d\n", (i % 16));
-		}
-		p->cb_printf ("s-%d\n", len);
-		break;
-	case 'A': // "pcA"
-		/* implemented in core because of disasm :( */
-		break;
-	case 'a': // "pca"
-		p->cb_printf ("shellcode:");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			if (!(i % 8)) {
-				p->cb_printf ("\n.byte ");
-			} else {
-				p->cb_printf (", ");
-			}
-			p->cb_printf ("0x%02x", buf[i]);
-		}
-		p->cb_printf ("\n.equ shellcode_len, %d\n", len);
-		break;
-	case 's': // "pcs"
-		p->cb_printf ("\"");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			p->cb_printf ("\\x%02x", buf[i]);
-		}
-		p->cb_printf ("\"\n");
-		break;
-	case 'S': // "pcS"
-	{
-		const int trunksize = 16;
-		for (i = 0; !p->interrupt && i < len; i++) {
-			if (!(i % trunksize)) {
-				p->cb_printf ("printf \"");
-			}
-			p->cb_printf ("\\%03o", buf[i]);
-			if ((i % trunksize) == (trunksize - 1)) {
-				p->cb_printf ("\" %s bin\n", (i <= trunksize)? ">": ">>");
-			}
-		}
-		if ((i % trunksize)) {
-			p->cb_printf ("\" %s bin\n", (i <= trunksize)? ">": ">>");
-		}
-	} break;
-	case 'J': { // "pcJ"
-		char *out = malloc (len * 3);
-		p->cb_printf ("var buffer = new Buffer(\"");
-		out[0] = 0;
-		r_base64_encode (out, buf, len);
-		p->cb_printf ("%s", out);
-		p->cb_printf ("\", 'base64');\n");
-		free (out);
-	} break;
-	case 'k': // "pck" kotlin
-		p->cb_printf ("val arr = byteArrayOfInts(");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("0x%x%s", buf[i], (i + 1 < len)? ",": "");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf (")\n");
-		break;
-	case 'z': // "pcz" // swift
-		p->cb_printf ("let byteArray : [UInt8] = [");
-
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("0x%x%s", buf[i], (i + 1 < len)? ", ": "");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf ("]\n");
-		break;
-	case 'r': // "pcr" // Rust
-		p->cb_printf ("let _: [u8; %d] = [\n", len);
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("0x%x%s", buf[i], (i + 1 < len)? ",": "");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf ("];\n");
-		break;
-	case 'o': // "pco" // Objective-C
-		p->cb_printf ("NSData *endMarker = [[NSData alloc] initWithBytes:{\n");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("0x%x%s", buf[i], (i + 1 < len)? ",": "");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf ("}];\n");
-		break;
-	case 'v': // "pcv" // JaVa
-		p->cb_printf ("byte[] ba = {");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("%d%s", buf[i], (i + 1 < len)? ",": "");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf ("};\n");
-		break;
-	case 'V': // "pcV" // vlang.io
-		p->cb_printf ("data := [ byte(%d),\n  ", buf[0]);
-		for (i = 1; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("%d%s", buf[i], (i + 1 < len)? ", ": "");
-			r_print_cursor (p, i, 1, 0);
-			if ((i %10) == 0) {
-				p->cb_printf ("\n  ");
-			}
-		}
-		p->cb_printf ("\n]\n");
-		break;
-	case 'y': // "pcy"
-		p->cb_printf ("$hex_%"PFMT64x" = {");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf (" %02x", buf[i] & 0xff);
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf (" }\n");
-		break;
-	case 'j': // "pcj"
-		p->cb_printf ("[");
-		for (i = 0; !p->interrupt && i < len; i++) {
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("%d%s", buf[i], (i + 1 < len)? ",": "");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf ("]\n");
-		break;
-	case 'P':
-	case 'p': // "pcp" "pcP"
-		p->cb_printf ("import struct\nbuf = struct.pack (\"%dB\", *[", len);
-		for (i = 0; !p->interrupt && i < len; i++) {
-			if (!(i % w)) {
-				p->cb_printf ("\n");
-			}
-			r_print_cursor (p, i, 1, 1);
-			p->cb_printf ("0x%02x%s", buf[i], (i + 1 < len)? ",": "])");
-			r_print_cursor (p, i, 1, 0);
-		}
-		p->cb_printf ("\n");
-		break;
-	case 'h': // "pch"
-		print_c_code (p, addr, buf, len, 2, p->cols / 2); // 9
-		break;
-	case 'w': // "pcw"
-		print_c_code (p, addr, buf, len, 4, p->cols / 3); // 6);
-		break;
-	case 'i': // "pci"
-		print_c_instructions (p, addr, buf, len, 1);
-		break;
-	case 'd': // "pcd"
-		print_c_code (p, addr, buf, len, 8, p->cols / 5); //3);
-		break;
-	default:
-		print_c_code (p, addr, buf, len, 1, p->cols / 1.5); // 12);
-		break;
-	}
-}
-
 static void print_json_string(RCore *core, const char* block, int len, const char* type) {
 	const char* section_name = r_core_get_section_name (core, core->offset);
 	if (section_name && strlen (section_name) < 1) {
@@ -5975,7 +5687,9 @@ l = use_blocksize;
 		}
 		break;
 	case 'c': // "pc"
-		if (l) {
+		if (input[1] == '?') {
+			r_core_cmd_help (core, help_msg_pc);
+		} else if (l) {
 			const ut8 *buf = core->block;
 			int i = 0;
 			int j = 0;
